% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_length.R
\name{age_length}
\alias{age_length}
\title{Uncertainty in age given length}
\usage{
age_length(highs, lows, L_a, CVlen)
}
\arguments{
\item{highs}{upper length bins}

\item{lows}{lower length bins}

\item{L_a}{Predicted length-at-age}

\item{CVlen}{CV of predicted length at age}
}
\value{
matrix - probability of being in a length bin given age
}
\description{
\code{age_length} Calculates probability of being in a length bin given age
}
\author{
M.B. Rudd
}
