% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeToLengthComp.R
\name{AgeToLengthComp}
\alias{AgeToLengthComp}
\title{Age-converted-to-length structure}
\usage{
AgeToLengthComp(lh, tyears, N_at, comp_sample)
}
\arguments{
\item{lh}{list of life history attributes, output of create_lh_list}

\item{tyears}{number of years of data}

\item{N_at}{matrix of numbers in the population at each age over time}

\item{comp_sample}{vector of number of individuals sampled each year (set as 1 for proportions)}
}
\value{
data frame - number of individuals in each length bin in each year
}
\description{
\code{AgeToLengthComp} Converts vulnerable numbers at age to length composition of catch
}
