% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AIC.R
\name{calc_AIC}
\alias{calc_AIC}
\title{Calculate AIC}
\usage{
calc_AIC(modpath_vec)
}
\arguments{
\item{modpath_vec}{path to model run}
}
\value{
matrix with AIC, AICc, deltaAIC, and deltaAICc (in that order by column) for each model (down the rows)
}
\description{
\code{calc_AIC} calculate AIC for model runs
}

