% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{Calc_derived_quants}
\alias{Calc_derived_quants}
\title{Calculate biological reference points}
\usage{
Calc_derived_quants(Obj)
}
\arguments{
\item{Obj,}{the fitted TMB object}
}
\value{
List, a tagged list of potentially useful benchmarks
}
\description{
\code{Calc_derived_quants} calculates derived quanties for status or productivity
}

