% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_equil_abund.R
\name{calc_equil_abund}
\alias{calc_equil_abund}
\title{Calculate abundance at equilibrium}
\usage{
calc_equil_abund(ages, S_a, M, F, R0)
}
\arguments{
\item{ages}{vector of ages}

\item{S_a}{selectivity-at-age}

\item{M}{natural mortality}

\item{F}{fishing mortality}

\item{R0}{recruitment}
}
\value{
vector of expected abundance at age
}
\description{
\code{calc_equil_abund} calculates expected abundance given constant fishing mortality
}
