% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_msy.R
\name{calc_msy}
\alias{calc_msy}
\title{Calculate MSY reference point}
\usage{
calc_msy(F, ages, S_a, M, R0, W_a, Mat_a)
}
\arguments{
\item{F}{fishing mortality rate}

\item{ages}{ages in the population}

\item{S_a}{selectivity-at-age}

\item{M}{natural mortality}

\item{R0}{equilibrium recruits}

\item{W_a}{weight-at-age}

\item{Mat_a}{maturity-at-age}
}
\value{
yield per recruit, in biomass
}
\description{
\code{calc_msy} calculates yield per recruit, when maximized across levels of F produces Fmsy
}
\author{
M.B. Rudd
}

