% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRE.R
\name{calcRE}
\alias{calcRE}
\title{Calculate relative error}
\usage{
calcRE(modpath_vec, itervec, value, yr, plot_nonconverge = FALSE,
  timeseries = FALSE)
}
\arguments{
\item{modpath_vec}{vector of directories to search for saved true and estimated population parameters}

\item{itervec}{vector of iterations to check for results}

\item{value}{options: "SPR" or "FFref" (FFref is defined as F/F40)}

\item{yr}{if timeseries==FALSE, specify the year for which to take the relative error}

\item{plot_nonconverge}{should the nonconverged parameter values be plotted? default FALSE}

\item{timeseries}{default: timeseries=FALSE, if timeseries=TRUE, takes the relative error for all years up to the specified year using the 'yr' argument}
}
\value{
list with Relative Error, Squared Error, Estimation Error, and vector flagging convergence issues
}
\description{
\code{calcRE} Calculates the relative error between true and estimated values of SPR and F/Fref
}
\details{
Only set up to run with a simulation study where there are multiple iterations of the model runs.
}

