% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{Check_Identifiable2}
\alias{Check_Identifiable2}
\title{Check for identifiability of fixed effects -- frm TMBhelpers (slightly adjusted by MBR)}
\usage{
Check_Identifiable2(obj)
}
\arguments{
\item{obj,}{The compiled TMB object}
}
\value{
A tagged list of the hessian and the message
}
\description{
\code{Check_Identifiable} calculates the matrix of second-derivatives of the marginal likelihood w.r.t. fixed effects, to see if any linear combinations are unidentifiable
}
\details{
Slight adjustment made in the RowMax calculation - when the eigenvector was a vector, the apply function was not working
}

