% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_sim.R
\name{copy_sim}
\alias{copy_sim}
\title{Copy simulated data from one data availability scenario to another}
\usage{
copy_sim(fromdir, fromcombos, todir, itervec, rewrite, res_dir)
}
\arguments{
\item{fromdir}{single directory from which to copy data}

\item{fromcombos}{data frame indicating the details of that single directory from which the data will be copied}

\item{todir}{list of possible directories to copy the data to. the function will filter and copy to the appropriate directories with the same sample size, life history, fishing mortality, and recruitment dynamics}

\item{itervec}{iterations of simulated data}

\item{rewrite}{if TRUE will overwrite the files, if FALSE will not overwrite the files if one already exists.}

\item{res_dir}{name for results directory to search for file (e.g. "equil" or "base" where specific models are written)}
}
\value{
saves files in new directory
}
\description{
\code{copy_sim} If two scenarios have the same population dynamics but different data availability going into the estimation model, will copy simulated population from one folder to another with the correct number of years of length composition data
}

