% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_cases.R
\name{cosine_distance_vector_to_matrix_rows}
\alias{cosine_distance_vector_to_matrix_rows}
\title{Compute distances between a dense vector and a sparse matrix}
\usage{
cosine_distance_vector_to_matrix_rows(vector, sparse_matrix)
}
\arguments{
\item{vector}{dense integer vector}

\item{sparse_matrix}{a sparse matrix of permutations}
}
\description{
Compute distances between a dense vector and a sparse matrix
}
