% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inputs.R
\name{create_inputs}
\alias{create_inputs}
\title{Create input parameters for TMB model}
\usage{
create_inputs(lh, input_data)
}
\arguments{
\item{lh}{tagged list of life history/starting value information}

\item{input_data}{tagged list of data inputs. Required: years = vector of years (true years or indices); LF = matrix of length frequency (years along rows and length bins along columns), obs_per_year = vector of sample size per year. Optional: I_t = vector of abundance index, named with years; C_t = vector of catch, named with years.}
}
\value{
List, a tagged list of potentially useful benchmarks
}
\description{
\code{create_inputs} Gets list of parameter inputs into the proper format
}
\details{
Specifically used to merge life history information with other model settings; flexibility to change parameter inputs for sensitivity analysis without changing the baseline life history information that was used to generate data in a simulation study, or carefully compiled for real=life application
}
\author{
M.B. Rudd
}
