% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{create_lh_list}
\alias{create_lh_list}
\title{Create hypothetical life histories for simulation model}
\usage{
create_lh_list(lh, param_adjust = FALSE, val = FALSE, selex, nlbins = 50)
}
\arguments{
\item{lh}{choose from different life histories, 1-4}

\item{param_adjust}{possibility of adjusting true parameter values - names of parameters, default=FALSE}

\item{val}{possibility of adjusting true parameter values - values for parameter names in "param_adjust" (must specify val if param_adjust is specified), default=FALSE}

\item{selex}{"asymptotic" selectivity or "dome"-shaped selectivity as the true selectivity in the operating model}

\item{nlbins}{specified number of length bins so that all life histories match (useful in simulation study)}
}
\value{
List, a tagged list of life history traits
}
\description{
\code{create_lh_list} Specifies 4 distinct life history types for simulation testing
}
\details{
Life histories used from Hordyk et al. 2015 ICES Journal of Marine Science, simulation test of LB-SPR method
}

