% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{data_avail_settings}
\alias{data_avail_settings}
\title{Settings for simulation testing of data availability scenarios}
\usage{
data_avail_settings(avail_set, ESS, simulation = TRUE)
}
\arguments{
\item{avail_set}{names of data availability scenarios, must match names in function, see details}

\item{ESS}{Effective sample size, default 1000}

\item{simulation}{default=TRUE, not set up for simulation=FALSE}
}
\value{
List, a tagged list of model settings
}
\description{
\code{data_avail_settings} Specifies 4 distinct life history types for simulation testing
}
\details{
avail_set: "Rich_LC"= 20 years length comp, index, and catch, high ESS; "Moderate_LC"=20 years length comp, index, and catch, lower ESS;  "Sample_LC"=20 years length comp, index, and catch, but only sampled every few years, and catch is reported at a rate of 20 percent; "Index_LC1"=20 years of abundance index, no catch data, 1 year of Length comp (can also specify "Index_LC10" for 10 years of length comp); "Catch_LC10"=20 years of catch data, 1 year of length comp (can also specify "Catch_LC10" for 10 years of length comp); "LC1" is 1 year of length comp only; can also specify LC2, 5, 10, and 20.
}

