% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{default_predict}
\alias{default_predict}
\title{Default function to perform the prediction}
\usage{
default_predict(data, model)
}
\arguments{
\item{data}{data to be explained (character vector)}

\item{model}{model to be explained}
}
\description{
Takes care of performing the prediction and adapting the format
}
