% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{default_tokenization}
\alias{default_tokenization}
\title{Default function to tokenize}
\usage{
default_tokenization(text)
}
\arguments{
\item{text}{text to tokenize as a character vector}
}
\description{
Use simple regex to tokenize a String.
}
