% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{.load_image_example}
\alias{.load_image_example}
\title{Load an example image explanation}
\usage{
.load_image_example()
}
\value{
A data.frame containing an example of a image explanation
}
\description{
This function is needed to cut down on package size. It reassembles the
explanation data.frame by attaching the image data to the saved data.frame
}
\keyword{internal}
