% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{.load_text_example}
\alias{.load_text_example}
\title{Load an example text explanation}
\usage{
.load_text_example()
}
\value{
A data.frame containing an example of a text explanation
}
\description{
Load an example text explanation
}
\keyword{internal}
