% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_input.R
\name{format_input}
\alias{format_input}
\title{TMB input formatting}
\usage{
format_input(input, data_avail, Fpen, SigRpen, SigRprior, LFdist, C_type,
  est_more, fix_more, f_startval_ft, rdev_startval_t, est_selex_f,
  vals_selex_ft, randomR, mirror, est_totalF, prop_f)
}
\arguments{
\item{input}{tagged list of LIME inputs. Output from create_inputs.}

\item{data_avail}{types of data included, must at least include LCX where X is the number of years of length composition data. May also include "Catch" or "Index" separated by underscore. For example, "LC10", "Catch_LC1", "Index_Catch_LC20".}

\item{Fpen}{penalty on fishing mortality 0= off, 1=on}

\item{SigRpen}{penalty on sigmaR, 0=off, 1=on}

\item{SigRprior}{vector with prior info for sigmaR penalty, first term is the mean and second term is the standard deviation}

\item{LFdist}{likelihood distribution for length composition data, default=0 for multinomial, alternate=1 for dirichlet-multinomial}

\item{C_type}{default=0, NO catch data available. Copt=1 means the catch is in numbers, Copt2 means the catch is in weight.}

\item{est_more}{list of variance parameters to estimate, must match parameter names: log_sigma_R, log_sigma_C, log_sigma_I, log_CV_L, log_sigma_F}

\item{fix_more}{default=FALSE - parameters are fixed depending on the data available. Can also list vector of parameter names to fix at their starting values (use param_adjust and val_adjust to set these adjustments)}

\item{f_startval_ft}{default=NULL and F starting values are at 0 for all years. Can also specify vector of F starting values for all years to be modeled (can start at truth for debugging).}

\item{rdev_startval_t}{default=NULL and Recruitment deviation starting values are at 0 for all years. Can also specify vector of recruitment deviation starting values for all years to be modeled (can start at truth for debugging)}

\item{est_selex_f}{default=TRUE to estimate selectivity parameters, can set to FALSE for all or multiple fleets}

\item{vals_selex_ft}{input selectivity-at-length (columns) by fleet (rows) - negative values in the first column indicate to estimate selectivity}

\item{randomR}{default = TRUE, estimate recruitment as a random effect; if FALSE, turn off random effect on recruitment (do not derive deviations)}

\item{mirror}{vector of parameter names to mirror between fleets}

\item{est_totalF}{TRUE estimate total F instead of by fleet}

\item{prop_f}{proportion of catch from each fleet}
}
\value{
List, a tagged list of Data, Parameters, Random, Map
}
\description{
\code{format_input} Formats data, parameters, random effects, and mapped parameters for TMB input
}
\author{
M.B. Rudd
}

