% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatData}
\alias{formatData}
\title{Generic function to format required LIME data}
\usage{
formatData(lfreq, lfreq_years, lbins, obs_per_year, index = NULL,
  index_years = NULL, catch = NULL, catch_years = NULL, meanlen = NULL,
  meanlen_years = NULL, model_years)
}
\arguments{
\item{lfreq}{matrix of length frequency data (number of fish in each length bin), with years along rows and bins along columns}

\item{lfreq_years}{vector of years for length frequency data}

\item{lbins}{length bins for length frequency data}

\item{obs_per_year}{vector wth sample size of length measurements each year}

\item{index}{vector of abundance index annually, if available. default=NULL}

\item{index_years}{vector of years for abundance index, if available. default=NULL}

\item{catch}{vector of catch annually, if available. default=NULL}

\item{catch_years}{vector of years for catch data, if available. default=NULL}

\item{meanlen}{vector of mean length annually, if available instead of length composition data. default=NULL}

\item{meanlen_years}{vector of years for mean length data, if available instead of length composition data. default=NULL}

\item{model_years}{number of years to be modeled - must be at least the continuous years for which data is available. can be longer. must be specified}
}
\value{
List, a tagged list of data for input into model
}
\description{
\code{formatData} Arguments specify the data and assumed values required for LIME, to get everything in one list
}
\details{
required output: I_t: index with each element named by year 1-x, C_t: catch with each element named 1-x, LF: length frequency with years 1-x labeled on the rows and length bin labeled on the columns, LFprop: proportions in each length bin, same dimensions as LF, years: actual years of data, years_i: index years 1-x, lbins: length bins, ML_t: mean length with each element named year 1-x, Nyears: number of years, Nyears_comp: number of years o f length composition data, obs_per_yr: effective sample size of length composition annually
}

