% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{FormatInput_LB}
\alias{FormatInput_LB}
\title{TMB input formatting}
\usage{
FormatInput_LB(Nyears, DataList, linf, vbk, t0, M, AgeMax, lbhighs, lbmids,
  Mat_a, lwa, lwb, log_sigma_C, log_sigma_I, log_CV_L, F1, SigmaR, qcoef, R0,
  S50, S95, model, RecDev_biasadj, site, Fpen, Dpen, Dprior, SigRpen, SigRprior,
  obs_per_yr, SigmaF, RecType, FType, LType, h, SelexTypeDesc, est_sigma, REML,
  estimate_same, start_f)
}
\arguments{
\item{Nyears}{total number of years to model}

\item{DataList}{list of data required for model}

\item{linf}{fixed life history parameter from input list}

\item{vbk}{fixed life history parameter from input list}

\item{t0}{fixed life history parameter from input list}

\item{M}{fixed life history parameter from input list}

\item{AgeMax}{fixed life history parameter from input list}

\item{lbhighs}{fixed length bins from input list}

\item{lbmids}{fixed length bins from input list}

\item{Mat_a}{fixed life history parameter from input list}

\item{lwa}{fixed life history parameter from input list}

\item{lwb}{fixed life history parameter from input list}

\item{log_sigma_C}{starting value from input list}

\item{log_sigma_I}{starting value from input list}

\item{log_CV_L}{starting value from input list}

\item{F1}{starting value from input list}

\item{SigmaR}{starting value from input list}

\item{qcoef}{starting value from input list}

\item{R0}{starting value from input list}

\item{S50}{starting value from input list}

\item{S95}{starting value from input list}

\item{model}{data type availability}

\item{RecDev_biasadj}{starting values for rec devs}

\item{site}{artifact of development: how many sites to estimate values?}

\item{Fpen}{penalty on fishing mortality 0= off, 1=on}

\item{Dpen}{penalty on terminal depletion 0= off, 1=on}

\item{Dprior}{prior info for depletion penalty}

\item{SigRpen}{penalty on sigmaR, 0=off, 1=on}

\item{SigRprior}{prior info for sigmaR penalty}

\item{obs_per_yr}{effective sample sizes for length comp (for each of total years)}

\item{SigmaF}{starting value from input list}

\item{RecType}{artifact of development: 0=random effects on recruitment}

\item{FType}{artifact of development: 0=fishing mortality fixed effects}

\item{LType}{artifact of development: 1= pooled length comp between sites}

\item{h}{starting value from input list}

\item{SelexTypeDesc}{artifact of development - asymptotic or dome-shaped selectivity}

\item{est_sigma}{list of variance parameters to estimate, must match parameter names}

\item{REML}{default FALSE}

\item{estimate_same}{TRUE=estimate least-common-denominator parameters for all data availability scenarios, FALSE=estimate parameters specific to data availability}

\item{start_f}{year (in numbers, not actual year) to start estimating fishing mortality (e.g. year 11 out of 20 to get estimates for last 10 years); the value of F in this year will be used as the estimate and SE for all previous years. 0=estimate all years.}
}
\value{
List, a tagged list of Data, Parameters, Random, Map
}
\description{
\code{FormatInput_LB} Formats data, parameters, random effects, and mapped parameters for TMB input
}

