% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data for simulation testing}
\usage{
generate_data(modpath, itervec, Fdynamics, Rdynamics, lh, pool = TRUE, Nyears,
  Nyears_comp, comp_sample, rewrite = TRUE, mismatch = FALSE, init_depl,
  derive_quants = FALSE, nburn = 50)
}
\arguments{
\item{modpath}{directory to save generated data}

\item{itervec}{number of iterations of data to generate}

\item{Fdynamics}{Specify name of pattern of fishing mortality dynamics, Constant, Endogenous, Ramp, Increasing, or None}

\item{Rdynamics}{Specify name of pattern of recruitment dynamics, Constant, Pulsed, Pulsed_up, or BH}

\item{lh}{list of life history information to feed to population simulation function, output from create_lh_list}

\item{pool}{if nseasons (in life history list) is greater than one, pool the generated data into annual time steps, or leave at the season-level? default=TRUE, FALSE will generate shorter time step life history info, mean length}

\item{Nyears}{number of years to simulate}

\item{Nyears_comp}{number of years to generate length composition data}

\item{comp_sample}{sample size of length composition data each year}

\item{rewrite}{TRUE will re-run OM and observation model. FALSE will skip if it's already written in directory.}

\item{mismatch}{default=FALSE, if TRUE, catch and index overlap with length comp only 1 year}

\item{init_depl}{default=0.4, can specify a different value or 2 values that indicate range from which to choose them}

\item{derive_quants}{default=FALSE (takes longer to run), can set to TRUE to output additional derived quantities.}

\item{nburn}{number of years burn-in for operating model, default=50}
}
\value{
print how many iterations were written into the model directory
}
\description{
\code{generate_data} Generates data from the operating model for use in simulation testing
}
\author{
M.B. Rudd
}

