% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{generateData}
\alias{generateData}
\title{Generate data for simulation testing}
\usage{
generateData(modpath, modname, itervec, spatial, Fdynamics, Rdynamics,
  LType = 1, plotML = FALSE, plotLF_compare = FALSE, plotLF = FALSE,
  selex = "asymptotic", write = TRUE, lh_list, data_avail_list,
  param_adjust = FALSE, val = FALSE, rewrite = TRUE)
}
\arguments{
\item{modpath}{directory to save generated data}

\item{modname}{name of model (to identify differences between different model runs)}

\item{itervec}{number of iterations of data to generate}

\item{spatial}{does asymptotic length vary by spatially? TRUE or FALSE}

\item{Fdynamics}{Specify name of pattern of fishing mortality dynamics, "Constant", "Endogenous", "Ramped", "Increasing", or "None"}

\item{Rdynamics}{Specify name of pattern of recruitment dynamics, "Constant", "Pulsed", "Pulsed_up", or "BH"}

\item{LType}{1 (default): pool site-specific length composition into 1 dataset; 0: keep length composition collected at different sites separate by site}

\item{plotML}{plot of generated mean length data, default=FALSE}

\item{plotLF_compare}{plot length comp pooled vs not pooled, default=FALSE}

\item{plotLF}{plot length composition at chosen LType, default=FALSE}

\item{selex}{"asymptotic" for asympotitc selectivity in length composition generation (currently only option programmed)}

\item{write}{write generated dataset? default=TRUE. FALSE helpful for plotting.}

\item{lh_list}{list of life history inputs}

\item{data_avail_list}{list of other model settings}

\item{param_adjust}{vector of names of parameters to adjust true value, default FALSE to include no parameter}

\item{val}{vector of values aligning with names in param_adjust to adjust value to, default FALSE to include no parameter}
}
\value{
print how many iterations were written into the model directory
}
\description{
\code{generateData} Generates data from the operating model for use in simulation testing
}

