% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_converged.R
\name{get_converged}
\alias{get_converged}
\title{Check convergence and re-run for common issues}
\usage{
get_converged(
  results,
  max_gradient = 0.001,
  saveFlagsDir = FALSE,
  saveFlagsName = FALSE
)
}
\arguments{
\item{results}{list of LIME results from `run_LIME`}

\item{max_gradient}{maximum gradient, default=0.001}

\item{saveFlagsDir}{directory to save flags for what was changed to get model converged}

\item{saveFlagsName}{name of model or node to save flags for what was changed or get model converged}
}
\value{
prints how many iterations were run in model directory
}
\description{
\code{get_converged} re-run LIME by pre-identified adjustments if not converged
}
\author{
M.B. Rudd
}
