% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_index_permutations}
\alias{get_index_permutations}
\title{Generate permutations of a vector of indexes.}
\usage{
get_index_permutations(original_document, number_permutations)
}
\arguments{
\item{original_document}{Index of words in a text document.}

\item{number_permutations}{number of permutation to generate.}
}
\value{
a list of permutations. The permutation sizes are not all the same.
}
\description{
First permutation is the original (full) document.
}
