% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFt.R
\name{getFt}
\alias{getFt}
\title{use newtons method to find instantaneous f
\code{getft} approximates f using newtons method based on catch, life history, and selectivity}
\usage{
getFt(ct, m, sa, wa, na)
}
\arguments{
\item{ct}{catch}

\item{m}{natural mortality}

\item{sa}{selectivity at age}

\item{wa}{weight at age}

\item{na}{abundance at age}
}
\value{
instantaneous fishing mortality rate f
}
\description{
use newtons method to find instantaneous f
\code{getft} approximates f using newtons method based on catch, life history, and selectivity
}
\author{
s.j.d. martell
}
