% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIME_fits.R
\name{LIME_fits}
\alias{LIME_fits}
\title{Plot LIME model fits}
\usage{
LIME_fits(Inputs, Report, Sdreport, obsData, save)
}
\arguments{
\item{Inputs}{list of inputs for model run}

\item{Report}{report file from model run}

\item{Sdreport}{sdreport file from model run}

\item{obsData}{observed data}

\item{save}{TRUE if you want to save the figure to directory, FALSE if you want to view figure}
}
\value{
figure of LIME model fits for several population parameters
}
\description{
\code{LIME_fits} create figure of LIME model fits for several population parameters
}

