% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_paths.R
\name{model_paths}
\alias{model_paths}
\title{Set up model paths}
\usage{
model_paths(res_dir, modcombos)
}
\arguments{
\item{res_dir}{directory to store results}

\item{modcombos}{data frame of all model combinations}
}
\value{
vector of directory paths
}
\description{
\code{model_paths} sets up directory paths for model combinations
}

