% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytheme.R
\name{mytheme}
\alias{mytheme}
\title{Theme for ggplot2}
\usage{
mytheme(base_size = 14, base_family = "")
}
\arguments{
\item{base_size}{default text size for figure, default=14}

\item{base_family}{font type for figure, default=""}
}
\value{
theme for use with figures from ggplot2
}
\description{
\code{mytheme} Preferred theme for ggplot2
}
\author{
Darcy Webber
}
