% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{plot_explanations}
\alias{plot_explanations}
\title{View HTML rendering of regular expression match.}
\usage{
plot_explanations(explanations)
}
\arguments{
\item{explanations}{object returned by the \code{lime.character} function.}
}
\description{
\code{plot_explanations} shows the first match;
}
\examples{
\dontrun{
# Explaining a model based on text data

library(text2vec)
library(lime)
library(xgboost)

data(train_sentences)
data(test_sentences)

get.matrix <- function(text) {
  it <- itoken(text, progressbar = FALSE)
  create_dtm(it, vectorizer = hash_vectorizer())
}

dtm_train = get.matrix(train_sentences$text)

bst <- xgb.train(list(max_depth = 7, eta = 0.1, objective = "binary:logistic",
                 eval_metric = "error", nthread = 1),
                 xgb.DMatrix(dtm_train, label = train_sentences$class.text == "OWNX"),
                 nrounds = 50)

r <- lime(test_sentences[5, text], bst, get.matrix, n_labels = 1,
     number_features_explain = 2, keep_word_position = FALSE)()

 print(r)

 plot_explanations(r)
}

}
