% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_output.R
\name{plot_output}
\alias{plot_output}
\title{plot LIME output}
\usage{
plot_output(all_years, lc_years, Inputs = NULL, Report = NULL,
  Sdreport = NULL, LBSPR = NULL, lh, true_years, True = NULL,
  plot = c("Fish", "Rec", "SPR", "ML", "SB", "Selex"))
}
\arguments{
\item{all_years}{vector of all years modeled}

\item{lc_years}{vector of years with length composition data}

\item{Inputs}{LIME input file}

\item{Report}{LIME report file}

\item{Sdreport}{LIME standard error file}

\item{LBSPR}{LBSPR results - must have pLF = probability of being harvested in a length bin; default NULL}

\item{lh}{life history list}

\item{true_years}{vector of true years (in case all_years and lc_years are 1:20 instead of 1998:2017)}

\item{True}{default=NULL, possible to specify true list from generated data if simulation}

\item{plot}{options for plotting include "Fish"=fishing mortality, "Rec"=recruitment (LIME only), "SPR"=spawning potential ratio, "ML"=mean length (including observed values; LIME only), "SB"=spawning biomass (LIME only), "Selex"=selectivity-at-length}
}
\value{
figure with length composition data and model fits if Report or LBSPR are specified
}
\description{
\code{plot_output} plot output from LIME or LB-SPR
}
\author{
M.B. Rudd
}

