% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{plotRE}
\alias{plotRE}
\title{Plot relative error}
\usage{
plotRE(modpath_vec, itervec, modnames = NULL, xaxt = "n", yaxt = "n",
  value, ylim = c(-1, 1.5), col.plot = "steelblue",
  col.line = "goldenrod", yr)
}
\arguments{
\item{modpath_vec}{directory to find model results}

\item{itervec}{number of iterations}

\item{modnames}{model names for x axis}

\item{xaxt}{default "n" to remove x axis, follows same rules as plot function}

\item{yaxt}{default "n" to remove y axis, follows same rules as plot function}

\item{value}{relative error for specified reference point, "SPR" or "FFref" (associated with F/F40)}

\item{ylim}{set y axis limits}

\item{col.plot}{color for boxplots}

\item{col.line}{color for line designating 0 relative error}

\item{yr}{year to plot relative error}
}
\value{
vector of directory paths
}
\description{
\code{plotRE} boxplots comparing relative error across several models
}

