% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{plotResults}
\alias{plotResults}
\title{model fits or kobe plots}
\usage{
plotResults(Data, Report, Sdreport, Derived_quants, flag_convergence, parameter,
  xaxt = TRUE, ylab = FALSE, simulation = TRUE)
}
\arguments{
\item{Data}{read RDS from model directory}

\item{Report}{read RDS from model directory}

\item{Sdreport}{read RDS from model directory}

\item{Derived_quants}{read RDS from model directory}

\item{flag_convergence}{TRUE- flag in directory that model didn't converge, FALSE = no flag in directory}

\item{parameter}{which parameter to plot results for}

\item{xaxt}{plot xaxis, TRUE or FALSE}

\item{ylab}{plot ylab, TRUE or FALSE}

\item{simulation}{get different items if these are application or simulation results}
}
\value{
displays plot
}
\description{
\code{plotResults} plot results for time series of estimated/derived parameters or kobe plots
}
\details{
possible values for parameter argument for model fits: "B" biomass, "N" abundance, "ML" mean length, "R" recruitment, "F" fishing mortality, "D" relative biomass, "C" catch, "I" abundance index, for kobe plot, "kobe" will show SPR compared with F/F30 and F/F40
}

