% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{print.letter}
\alias{print.letter}
\title{Write letters for multi-panel figures}
\usage{
\method{print}{letter}(label = "(a)", xy = c(0.1, 0.925), ...)
}
\arguments{
\item{label}{label to plot on multi-panel figure}

\item{xy}{setting location of label}
}
\value{
adds text to figure
}
\description{
\code{plotResults} writes letters on multi-panel figures
}
\author{
Trevor A. Branch
}

