% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sdreport.R
\name{read_sdreport}
\alias{read_sdreport}
\title{read TMB Sdreport}
\usage{
read_sdreport(InputMat, log = TRUE)
}
\arguments{
\item{InputMat}{summary(Sdreport) which rownames have specific parameter name}

\item{log}{is the parameter in logspace?}
}
\value{
vector with lower and upper 95 confidence intervals. Upper intervals are reversed for ease plotting using polygon function.
}
\description{
\code{read_sdreport} read summary(Sdreport) to calculate confidence intervals for a parameter
}
