% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_text.R
\name{render_text_explanations}
\alias{render_text_explanations}
\title{Shiny widget render}
\usage{
render_text_explanations(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates an HTML widget}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{\link[=quote]{quote()}})? This
is useful if you want to save an expression in a variable.}
}
\value{
A render function that enables the use of the widget within Shiny applications.
}
\description{
Render the text explanations in Shiny application.
}
