% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rowSumsSq}
\alias{rowSumsSq}
\title{https://codereview.stackexchange.com/questions/159396/cosine-similarity-of-one-vector-with-many
Modified to manage sparse matrix with Eigen}
\usage{
rowSumsSq(sparse_matrix)
}
\arguments{
\item{sparse_matrix}{the matrix to compute distance on.}
}
\description{
https://codereview.stackexchange.com/questions/159396/cosine-similarity-of-one-vector-with-many
Modified to manage sparse matrix with Eigen
}
