% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_LIME.R
\name{run_LIME}
\alias{run_LIME}
\title{run LIME model - simulation mode}
\usage{
run_LIME(
  modpath,
  input,
  data_avail,
  Fpen = 1,
  SigRpen = 1,
  SigRprior = c(0.737, 0.3),
  LFdist = 1,
  C_type = 0,
  est_more = FALSE,
  fix_more = FALSE,
  est_F_ft = TRUE,
  f_startval_ft = NULL,
  rdev_startval_t = NULL,
  est_selex_f = TRUE,
  vals_selex_ft = -1,
  est_rdev_t = TRUE,
  newtonsteps = 3,
  F_up = 10,
  S50_up = NULL,
  derive_quants = FALSE,
  itervec = NULL,
  simulation = FALSE,
  rewrite = TRUE,
  mirror = NULL,
  est_totalF = FALSE,
  prop_f = 1
)
}
\arguments{
\item{modpath}{model directory}

\item{input}{tagged list of LIME inputs. Output from create_inputs.}

\item{data_avail}{types of data included, must at least include LCX where X is the number of years of length composition data. May also include "Catch" or "Index" separated by underscore. For example, "LC10", "Catch_LC1", "Index_Catch_LC20".}

\item{Fpen}{penalty on fishing mortality 0= off, 1=on}

\item{SigRpen}{penalty on sigmaR, 0=off, 1=on}

\item{SigRprior}{vector with prior info for sigmaR penalty, first term is the mean and second term is the standard deviation}

\item{LFdist}{likelihood distribution for length composition data, default=0 for multinomial, alternate=1 for dirichlet-multinomial}

\item{C_type}{default=0, NO catch data available. Copt=1 means the catch is in numbers, Copt2 means the catch is in weight.}

\item{est_more}{list of variance parameters to estimate, must match parameter names: log_sigma_R, log_sigma_C, log_sigma_I, log_CV_L, log_sigma_F}

\item{fix_more}{default=FALSE - parameters are fixed depending on the data available. Can also list vector of parameter names to fix at their starting values (use param_adjust and val_adjust to set these adjustments)}

\item{est_F_ft}{default=TRUE, otherwise 0 for off and 1 for on in matrix that matches fleets in rows and years in columns}

\item{f_startval_ft}{default=NULL and F starting values are at 0 for all years. Can also specify vector of F starting values for all years to be modeled (can start at truth for debugging).}

\item{rdev_startval_t}{default=NULL and Recruitment deviation starting values are at 0 for all years. Can also specify vector of recruitment deviation starting values for all years to be modeled (can start at truth for debugging)}

\item{est_selex_f}{default=TRUE to estimate selectivity parameters, can set to FALSE for all or multiple fleets}

\item{vals_selex_ft}{input selectivity-at-length (columns) by fleet (rows) - negative values in the first column indicate to estimate selectivity}

\item{est_rdev_t}{default=TRUE to estimate recruitment deviations, or specify vector with 0 to turn off deviations in a specific year and 1 to keep them on}

\item{newtonsteps}{number of extra newton steps to take after optimization; FALSE to turn off}

\item{F_up}{upper bound of fishing mortality estimate; default=10}

\item{S50_up}{upper bound of length at 50 percent selectivity; default=NULL}

\item{derive_quants}{if TRUE, derive MSY-related reference points, default=FALSE}

\item{itervec}{number of datasets to generate in a simulation study. default=NULL for real stock assessment application.}

\item{simulation}{is this a simulation? default FALSE means you are using real data (can set itervec=NULL)}

\item{rewrite}{default=TRUE; if results already exist in the directory, should we rewrite them? TRUE or FALSE}

\item{mirror}{vector of parameter names to mirror between fleets}

\item{est_totalF}{TRUE estimate total F instead of by fleet}

\item{prop_f}{proportion of catch from each fleet}
}
\value{
prints how many iterations were run in model directory
}
\description{
\code{run_LIME} run length-based integrated mixed-effects model with generated data
}
\author{
M.B. Rudd
}
