% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stacking.R
\name{run_stacking}
\alias{run_stacking}
\title{Run stacking algorithm}
\usage{
run_stacking(modpath, iter = NULL, lh, model, nodes, dim,
  input_data = NULL)
}
\arguments{
\item{modpath, }{the directory to save results}

\item{iter, }{the iteration for a simulation study, default=NULL for an assessment}

\item{lh, }{the life history list output from LIME::create_lh_list}

\item{model, }{"LIME" or "LBSPR"}

\item{nodes, }{data frame of nodes mapped to life history parameters, number of rows = number of nodes, number of columns = life history parameter K, Loo, M, Lm}

\item{dim, }{"1D" to label 1-d stacking, "2D" to label 2-d stacking}

\item{input_data, }{default=NULL for simulation study, data list following rules by LIME for regular assessment}
}
\value{
List, a tagged list of potentially useful benchmarks
}
\description{
\code{run_stacking} runs stacking algorithm for length-based stock assessment methods given quadrature nodes
}
\author{
M.B. Rudd
}
