% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{runAssessment}
\alias{runAssessment}
\title{run LIME model - application mode}
\usage{
runAssessment(modpath, DataList, version = "LIME", lh_inputs, data_avail,
  est_sigma, biascorrect = TRUE, sensitivity_inputs = NULL,
  sensitivity_ESS = NULL, REML = FALSE)
}
\arguments{
\item{modpath}{model directory}

\item{DataList}{data list}

\item{version}{cpp model version name}

\item{lh_inputs}{life history inputs}

\item{data_avail}{other settings for data availability}

\item{est_sigma}{which variance parameters to estimate, match parameter names}

\item{biascorrect}{bias correction for recruitment deviatiosn on (TRUE) or off (FALSE)}

\item{sensitivity_inputs}{(in development) named list (parameters) with matrix of 2 rows (low, high) and # of columns for 'life histories' (artifact of testing multiple life histories in the simulation, would only have 1 column for an assessment)}

\item{sensitivity_ESS}{(in development) will do sensitivity analysis for effective sample size of length composition}

\item{REML}{default off (FALSE)}
}
\value{
displays plot
}
\description{
\code{runAssessment} run length-based integrated mixed-effects model with real data
}
\details{
removes iterations associated with runModel for simulation *** not updated - runModel is the most updated. Need to edit code and edit runModel to also run with real data, and delete this function.
}

