% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{runLBSPR}
\alias{runLBSPR}
\title{run LBSPR within LIME framework}
\usage{
runLBSPR(Nyears_comp, inits, iterpath, DataList, species)
}
\arguments{
\item{Nyears_comp}{number of years of length composition data}

\item{inits}{initial values for parameters}

\item{iterpath}{directories to run model}

\item{DataList}{list of real or generated data}

\item{species}{name of species (will set as generic within LBSPR package if not specified)}
}
\value{
list of LBSPR output
}
\description{
\code{runLBSPR} sets up LBSPR to run within LIME framework (adjusting for inputs, structure, etc.)
}
\details{
Must load LBSPR package: devtools::install_github("AdrianHordyk/LBSPR")
}

