% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{runModel}
\alias{runModel}
\title{run LIME model - simulation mode}
\usage{
runModel(modpath, itervec, estimate_same = FALSE, REML = FALSE, est_sigma,
  biascorrect = TRUE, data_avail, lh_list, sensitivity_inputs = NULL,
  sensitivity_ESS = NULL, rewrite, start_f, simulation = TRUE)
}
\arguments{
\item{modpath}{model directory}

\item{itervec}{number of iterations of data to generate}

\item{estimate_same}{TRUE=estimate least-common-denominator parameters for all data availability scenarios, FALSE=estimate parameters specific to data availability}

\item{REML}{default off (FALSE)}

\item{est_sigma}{which variance parameters to estimate, match parameter names}

\item{biascorrect}{bias correction for recruitment deviatiosn on (TRUE) or off (FALSE)}

\item{data_avail}{other settings for data availability}

\item{lh_list}{list of life history information}

\item{sensitivity_inputs}{(in development) named list (parameters) with matrix of 2 rows (low, high) and # of columns for 'life histories' (artifact of testing multiple life histories in the simulation, would only have 1 column for an assessment)}

\item{sensitivity_ESS}{(in development) will do sensitivity analysis for effective sample size of length composition}

\item{rewrite}{if results already exist in the directory, should we rewrite them? TRUE or FALSE}

\item{start_f}{year (in numbers, not actual year) to start estimating fishing mortality (e.g. year 11 out of 20 to get estimates for last 10 years); the value of F in this year will be used as the estimate and SE for all previous years. 0=estimate all years.}

\item{simulation}{is this a simulation? default TRUE, FALSE means you are using real data (no need for iterations or multiple life history inputs)}
}
\value{
prints how many iterations were run in model directory
}
\description{
\code{runModel} run length-based integrated mixed-effects model with generated data
}
\details{
need to adjust to run with real data
}

