% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runstack.R
\name{runstack}
\alias{runstack}
\title{Predictive stacking}
\usage{
runstack(savedir, iter, seed, lh, nodes, param, mean, cov, modname, input_data,
  Fscenario, rewrite, binwidth = 1)
}
\arguments{
\item{savedir}{directory to save results}

\item{iter}{iteration of generated data}

\item{seed}{set seed}

\item{lh}{life history list, with elements contained as output from LIME::create_lh_list}

\item{nodes}{matrix of nodes where each column is a different parameter and each row is a value from a distribution}

\item{param}{parameters (column names for nodes)}

\item{mean}{mean of each parameter value}

\item{cov}{covariance matrix across parameters}

\item{modname}{model name to save in directory}

\item{input_data}{for LIME use with real data (not simulated)}

\item{Fscenario}{fishing mortality scenario to generate data}

\item{rewrite}{rewrite results?}

\item{binwidth}{default=1cm, can change}
}
\value{
prints how many iterations were run in model directory
}
\description{
\code{runstack} run predicted stacking for life history parameters using LIME
}
\author{
M.B. Rudd
}

