% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_pop}
\alias{sim_pop}
\title{Operating model}
\usage{
sim_pop(lh, Nyears, pool, Fdynamics, Rdynamics, Nyears_comp, comp_sample,
  init_depl, nburn, seed, modname, mismatch, sample_type = "catch")
}
\arguments{
\item{lh}{list of life history information, from create_lh_list}

\item{Nyears}{number of years to simulate}

\item{pool}{if nseasons (in life history list) is greater than one, pool the generated data into annual time steps, or leave at the season-level? FALSE will generate shorter time step life history info, mean length}

\item{Fdynamics}{Specify name of pattern of fishing mortality dynamics, Constant, Endogenous, Ramp, Increasing, or None}

\item{Rdynamics}{Specify name of pattern of recruitment dynamics, Constant, Pulsed, Pulsed_up, or BH}

\item{Nyears_comp}{number of years of length composition data}

\item{comp_sample}{sample size of length composition data annually}

\item{init_depl}{initial depletion; if FALSE, will use F1 from lh list}

\item{nburn}{number of years of burn-in for operating model}

\item{seed}{set seed for generating stochastic time series}

\item{modname}{save model name for true dynamics in named list output}

\item{mismatch}{if TRUE, catch and index overlap with length comp only 1 year}

\item{sample_type}{a character vector specifying if the length comps are sampled from the 'catch' (default) or from the population}
}
\value{
named list of attributes of true population/data
}
\description{
\code{sim_pop} Age-converted-to-length-based operating model specifying true population dynamics
}
