% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{SimData_LB}
\alias{SimData_LB}
\title{Operating model}
\usage{
SimData_LB(Nyears, AgeMax, SigmaR, M, F1, S_a, h, qcoef, Frate, Fequil, SigmaF,
  Fdynamics, Rdynamics, R0, Fmax, CVlen, mids, highs, lows, W_a, L_a, linf, vbk,
  Mat_a, Amat, comp_sample, Nyears_comp, alt_yrs = FALSE, sample = FALSE,
  nburn, seed, modname)
}
\arguments{
\item{Nyears}{number of years to simulate}

\item{AgeMax}{maximum age in population}

\item{SigmaR}{Specified level of recruitment variation (as a standard deviation)}

\item{M}{natural mortality rate}

\item{F1}{Fishing mortality level in first year}

\item{S_a}{vector of selectivity at age}

\item{h}{steepness}

\item{qcoef}{catchability coefficient for abundance index}

\item{Frate}{parameter in Endogenous fishing mortality, see Thorson et al. 2014 CJFAS effort dynamics}

\item{Fequil}{parameter in Endogenous fishing mortality, see Thorson et al. 2014 CJFAS effort dynamics}

\item{SigmaF}{specified level of variation in fishing mortality annually (as a standard deviation)}

\item{Fdynamics}{Specify name of pattern of fishing mortality dynamics, "Constant", "Endogenous", "Ramped", "Increasing", or "None"}

\item{Rdynamics}{Specify name of pattern of recruitment dynamics, "Constant", "Pulsed", "Pulsed_up", or "BH"}

\item{R0}{equilibrium, initial recruitment}

\item{Fmax}{maximum possible fishing mortality}

\item{CVlen}{variation in growth at age as a coefficient of variation}

\item{mids}{midpoint of length bins}

\item{highs}{high end of length bins}

\item{lows}{low end of length bins}

\item{W_a}{weight at age}

\item{L_a}{length-at-age}

\item{linf}{asymptotic length}

\item{vbk}{Brody growth coefficient}

\item{Mat_a}{maturity at age}

\item{Amat}{Age at 50% maturity}

\item{comp_sample}{number of indiviuals sampled for lenght composition}

\item{Nyears_comp}{number of years of length composition data}

\item{alt_yrs}{only some years sampled for data? TRUE or FALSE (default FALSE)}

\item{sample}{only a sample of catch is reported? default FALSE}

\item{nburn}{number of years of burn-in for operating model}

\item{seed}{set seed for generating stochastic time series}

\item{modname}{save model name for true dynamics in named list output}
}
\value{
named list of attributes of true population/data
}
\description{
\code{SimData_LB} Age-converted-to-length-based operating model specifying true population dynamics
}

