% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{spatialgrowth_sim}
\alias{spatialgrowth_sim}
\title{Simulate spatial variation in growth}
\usage{
spatialgrowth_sim(n_i, Scale = 2, Sigma2 = 1, SD_spatial = 0.1, linf,
  beta_y = 0.02)
}
\arguments{
\item{n_i}{number of sites}

\item{Scale}{Gaussian scale parameter, default 2}

\item{Sigma2}{variance in asymptotic length across sites, default 1}

\item{SD_spatial}{Gaussian variation (standard deviation), default 0.1}

\item{linf}{average asymptotic length across sites}

\item{beta_y}{trend in asymptotic length over sites, default 0.02}
}
\value{
data frame of asymptotic length at each site
}
\description{
\code{spatialgrowth_sim} simulate 1D Spatial variation in asymptotic length across sites
}

