% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stop_words_sentences}
\alias{stop_words_sentences}
\title{Stop words list}
\format{Character vector of stop words}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/}
}
\usage{
stop_words_sentences
}
\description{
List of words that can be safely removed from sentences.
}
\keyword{datasets}
