% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_text.R
\name{plot_text_explanations}
\alias{plot_text_explanations}
\alias{text_explanations_output}
\alias{render_text_explanations}
\title{Plot text explanations}
\usage{
plot_text_explanations(explanations)

text_explanations_output(outputId, width = "100\%", height = "400px")

render_text_explanations(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{explanations}{object returned by the \link{lime.character} function.}

\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit or a number, which will be coerced to a string and have "px" appended.}

\item{expr}{An expression that generates an HTML widget}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{\link[=quote]{quote()}})? This
is useful if you want to save an expression in a variable.}
}
\value{
An output function that enables the use of the widget within Shiny applications.

A render function that enables the use of the widget within Shiny applications.
}
\description{
Highlight words which explains a prediction.

Create an output to insert text explanation plot in Shiny application.

Render the text explanations in Shiny application.
}
\examples{
\dontrun{
# Explaining a model based on text data

# Purpose is to classify sentences from scientific publications
# and find those where the team writes about their own work
# (category OWNX in the provided dataset).

library(lime)
library(text2vec)
library(xgboost)

data(train_sentences)
data(test_sentences)

get_matrix <- function(text) {
  it <- itoken(text, progressbar = FALSE)
  create_dtm(it, vectorizer = hash_vectorizer())
}

dtm_train = get_matrix(train_sentences$text)

xgb_model <- xgb.train(list(max_depth = 7, eta = 0.1, objective = "binary:logistic",
                 eval_metric = "error", nthread = 1),
                 xgb.DMatrix(dtm_train, label = train_sentences$class.text == "OWNX"),
                 nrounds = 50)

sentences <- head(test_sentences[test_sentences$class.text == "OWNX", "text"], 5)
explainer <- lime(sentences, xgb_model, get_matrix)
explanations <- explain(sentences, explainer, n_labels = 1, n_features = 2)

# We can see that many explanations are based
# on the presence of the word `we` in the sentences
# which makes sense regarding the task.
print(explanations)

# We can also print the explanations to see the selected words in their context.
plot_text_explanations(explanations)
}

}
