% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_text.R
\name{text_explanations_output}
\alias{text_explanations_output}
\title{Shiny widget output}
\usage{
text_explanations_output(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit or a number, which will be coerced to a string and have "px" appended.}
}
\value{
An output function that enables the use of the widget within Shiny applications.
}
\description{
Create an output to insert text explanation plot in Shiny application.
}
