% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{AgeToLengthComp}
\alias{AgeToLengthComp}
\title{Age-converted-to-length structure}
\usage{
AgeToLengthComp(L_a, CVlen, highs, lows, tyears, N_at, S_a, comp_sample)
}
\arguments{
\item{L_a}{vector, growth curve: predicted length at each age}

\item{CVlen}{coefficient of variation for growth curve}

\item{highs}{vector of upper end of length bins}

\item{lows}{vector of lower end of length bins}

\item{tyears}{number of years of data}

\item{N_at}{matrix of numbers in the population at each age over time}

\item{S_a}{vector of selectivity at age}

\item{comp_sample}{vector of number of individuals sampled each year (set as 1 for proportions)}
}
\value{
data frame - number of individuals in each length bin in each year
}
\description{
\code{AgeToLengthComp} Converts vulnerable numbers at age to length composition of catch
}

