% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeToLengthComp.R
\name{AgeToLengthComp}
\alias{AgeToLengthComp}
\title{Age-converted-to-length structure for compositional data}
\usage{
AgeToLengthComp(lh, S_a, tyears, N_at, comp_sample,
  sample_type = "catch")
}
\arguments{
\item{lh}{list of life history attributes, output of create_lh_list}

\item{S_a}{flexiblity to set selectivity-at-age by fleet}

\item{tyears}{number of years of data}

\item{N_at}{matrix of numbers in the population at each age over time}

\item{comp_sample}{vector of number of individuals sampled each year (set as 1 for proportions)}

\item{sample_type}{a character vector specifying if the length comps are sampled from the 'catch' (default) or from the population}
}
\value{
data frame - number of individuals in each length bin in each year
}
\description{
\code{AgeToLengthComp} Converts vulnerable numbers at age to length composition of catch
}
\author{
M.B. Rudd
}
