% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_derived_quants.R
\name{calc_derived_quants}
\alias{calc_derived_quants}
\title{Calculate biological reference points}
\usage{
calc_derived_quants(Obj)
}
\arguments{
\item{Obj, }{the fitted TMB object}
}
\value{
List, a tagged list of potentially useful benchmarks
}
\description{
\code{calc_derived_quants} calculates derived quanties for status or productivity
}

