% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_derived_quants.R
\name{calc_derived_quants}
\alias{calc_derived_quants}
\title{Calculate biological reference points}
\usage{
calc_derived_quants(Obj, lh)
}
\arguments{
\item{Obj, }{the fitted TMB object}

\item{lh, }{list of life history information}
}
\value{
List, a tagged list of potentially useful benchmarks
}
\description{
\code{calc_derived_quants} calculates derived quanties for status or productivity
}
\author{
M.B. Rudd
}

