% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ref.R
\name{calc_ref}
\alias{calc_ref}
\title{Per recruit calculation and spawning potential ratio}
\usage{
calc_ref(ages, Mat_a, W_a, M, F, S_fa, ref = FALSE, fleet_prop = 1,
  type = "SPR")
}
\arguments{
\item{ages}{vector of ages}

\item{Mat_a}{from report file / true file for simulation}

\item{W_a}{from report file / true file for simulation}

\item{M}{from report file / true file for simulation}

\item{F}{proposed F to calculate reference point, if S_fa has more than 1 fleet, requires F as a vector, one for each fleet}

\item{S_fa}{selectivity by fleet by age}

\item{ref}{FALSE outputs SPR, ref= a value between 0 and 1 can be used with uniroot to find the F at which SPR=ref}

\item{fleet_prop}{fleet proportions}

\item{type}{option spawning biomass per recruit (SPR) or biomass per recruit ("Depletion")}
}
\value{
List, a tagged list of potentially useful benchmarks
}
\description{
\code{calc_ref} Calculates derived spawning potential ratio: lifetime total egg production in fished:unfished states
}
\details{
Use this function with uniroot to find the value of F that results in SPR matching the specified reference value (e.g. 0.30 to find F30)
}
\author{
M.B. Rudd
}
