% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{calc_ref}
\alias{calc_ref}
\title{Per recruit calculation and spawning potential ratio}
\usage{
calc_ref(Mat_a, W_a, M, S_a, F, ref = FALSE)
}
\arguments{
\item{Mat_a}{from report file / true file for simulation}

\item{W_a}{from report file / true file for simulation}

\item{M}{from report file / true file for simulation}

\item{S_a}{from report file / true file for simulation}

\item{F}{typically terminal estimated/true F, can be any year}

\item{ref}{FALSE outputs SPR, ref= a value between 0 and 1 can be used with uniroot to find the F at which SPR=ref}
}
\value{
List, a tagged list of potentially useful benchmarks
}
\description{
\code{calc_ref} Calculates derived spawning potential ratio: lifetime total egg production in fished:unfished states
}
\details{
Use this function with uniroot to find the value of F that results in SPR matching the specified reference value (e.g. 0.30 to find F30)
}

