% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{calcRE}
\alias{calcRE}
\title{Calculate relative error}
\usage{
calcRE(modpath_vec, iter, value, yr, timeseries = FALSE)
}
\arguments{
\item{modpath_vec}{vector of directories to search for saved true and estimated population parameters}

\item{iter}{which iteration? only 1 at a time}

\item{value}{options: "SPR" or "FFref" (FFref is defined as F/F40%)}

\item{yr}{if timeseries==FALSE, specify the year for which to take the relative error}

\item{timeseries}{default: timeseries=FALSE, if timeseries=TRUE, takes the relative error for all years up to the specified year using the 'yr' argument}
}
\value{
Matrix with relative error for single iteration across all models
}
\description{
\code{calcRE} Calculates the relative error between true and estimated values of SPR and F/Fref
}
\details{
repeat over all iterations outside of this function to get full distribution of relative errors in the reference points for the simulation study.
}

