% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{choose_lh_list}
\alias{choose_lh_list}
\title{Saved life history values}
\usage{
choose_lh_list(species, selex)
}
\arguments{
\item{species}{species code name}

\item{selex}{"asymptotic"= assume asymptotic selectivity, "dome"= assume dome-shaped selectivity}
}
\value{
List, a tagged list of life history/starting value information
}
\description{
\code{choose_lh_list} Fixed lists of life history/starting value information for pre-loaded species
}

