% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inputs.R
\name{create_inputs}
\alias{create_inputs}
\title{Create input parameters for TMB model}
\usage{
create_inputs(param, val, lh_list, data_avail_list)
}
\arguments{
\item{param}{parameter name to adjust (sensitivity analysis)}

\item{val}{value of parameter name to adjust (sensitivity analysis)}

\item{lh_list}{tagged list of life history/starting value information}

\item{data_avail_list}{artifact from sensitivity analysis, adjusts some information on sample size, years, etc. for different data-availability scenarios}
}
\value{
List, a tagged list of potentially useful benchmarks
}
\description{
\code{create_inputs} Gets list of parameter inputs into the proper format
}
\details{
Specifically used to merge life history information with other model settings; flexibility to change parameter inputs for sensitivity analysis without changing the baseline life history information that was used to generate data in a simulation study, or carefully compiled for real=life application
}

