% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lh_list.R
\name{create_lh_list}
\alias{create_lh_list}
\title{Create new life history list}
\usage{
create_lh_list(vbk, linf, lwa, lwb, S50, M50, selex_input = "length",
  maturity_input = "length", binwidth = 1, t0 = -0.01, CVlen = 0.1,
  SigmaC = 0.2, SigmaI = 0.2, SigmaR = 0.6, SigmaF = 0.3, R0 = 1,
  h = 1, qcoef = 1e-05, M = NULL, F1 = 0.2, Fequil = 0.2,
  Frate = 0.2, Fmax = 0.7, start_ages = 0, rho = 0, Mat0 = TRUE,
  Sel0 = TRUE, theta = 10)
}
\arguments{
\item{vbk}{von Bertalanffy k Brody growth coefficient}

\item{linf}{von Bertalanffy Linf asymptotic length}

\item{lwa}{length-weight scaling parameter}

\item{lwb}{length-weight allometric parameter}

\item{S50}{starting value for age or length at 50 percent selectivity (will be estimated in LIME method)}

\item{M50}{age or length at 50 percent maturity}

\item{selex_input}{specify whether argument S50 is an age or a length (default length)}

\item{maturity_input}{specify whether argument M50 is an age or a length (default length)}

\item{binwidth}{width of length bins (default = 1)}

\item{t0}{theoretical age at length=0 (default = -0.01); avoid fixing to zero due to some issues with the first age/length bin}

\item{CVlen}{CV of the growth curve (default = 0.1)}

\item{SigmaC}{standard deviation - observation error of catch data (default = 0.2)}

\item{SigmaI}{standard deviation - observation error of index data (default = 0.2)}

\item{SigmaR}{standard deviation - process error for recruitment time series (default = 0.6 -- starting value, will be estimated)}

\item{SigmaF}{standard deviation - process error for fishing mortality time series (default = 0.3)}

\item{R0}{equilibrium recruitment (default = 1); when no information on scale is available, will estimate relative deviations around equilibrium 1}

\item{h}{steepness parameter (default = 1)}

\item{qcoef}{starting value for catchability coefficient (when index data is available, default = 1e-5)}

\item{M}{value for natural mortality if there has been a study (default = NULL, calculated internally from vbk)}

\item{F1}{starting value for initial fishing mortality. Default = 0.2, do not start at zero because this is used to set the initial values for estimating annual fishing mortality in log space, thus setting to zero would cause an error.}

\item{Fequil}{equilibrium fishing mortality rate (used for simulation; default=0.2)}

\item{Frate}{parameter used to simulate fishing moratality time series (default=NULL)}

\item{Fmax}{maximum F used in simulation (default=NULL)}

\item{start_ages}{age to start (either 0 or 1; default = 0)}

\item{rho}{first-order autocorrelation in recruitment residuals parameter, default=0 (recruitment not autocorrelated)}

\item{Mat0}{default=TRUE, maturity at age zero fixed to 1e-20, if FALSE, will calculate based on logistic curve}

\item{Sel0}{default=TRUE, selectivity at age zero fixed to 1e-20, if FALSE, will calculate based on logistic curve}

\item{theta}{dirichlet-multinomial parameter related to effective sample size. default to 10, will not be used if length frequency distribution LFdist is set to multinomial (0). Only used if distribution is dirichlet-multinomial (LFdist=1)}
}
\value{
List, a tagged list of life history traits
}
\description{
\code{create_lh_list} Creates list of life history information
}

