% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{default_tokenize}
\alias{default_tokenize}
\title{Default function to tokenize}
\usage{
default_tokenize(text)
}
\arguments{
\item{text}{text to tokenize as a \code{\link{character}} vector}
}
\description{
Use simple regex to tokenize a \code{\link{character}} vector. To be used with \code{\link{lime.character}}.
}
