% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_input.R
\name{format_input}
\alias{format_input}
\title{TMB input formatting}
\usage{
format_input(input, data_avail, Fpen, SigRpen, SigRprior, est_sigma, REML,
  fix_f, f_startval, fix_param = FALSE, C_opt = 0, Sel0, LFdist)
}
\arguments{
\item{input}{tagged list of LIME inputs. Output from create_inputs.}

\item{data_avail}{types of data included, must at least include LCX where X is the number of years of length composition data. May also include "Catch" or "Index" separated by underscore. For example, "LC10", "Catch_LC1", "Index_Catch_LC20".}

\item{Fpen}{penalty on fishing mortality 0= off, 1=on}

\item{SigRpen}{penalty on sigmaR, 0=off, 1=on}

\item{SigRprior}{vector with prior info for sigmaR penalty, first term is the mean and second term is the standard deviation}

\item{est_sigma}{list of variance parameters to estimate, must match parameter names: log_sigma_R, log_sigma_C, log_sigma_I, log_CV_L, log_sigma_F}

\item{REML}{FALSE==off, TRUE==on}

\item{fix_f}{year (by index, not actual year) to start estimating fishing mortality (e.g. year 11 out of 20 to get estimates for last 10 years); the value of F in this year will be used as the estimate and SE for all previous years. 0=estimate all years.}

\item{f_startval}{default=NULL and F starting values are at 0 for all years. Can also specify vector of F starting values for all years to be modeled (can start at truth for debugging).}

\item{fix_param}{default=FALSE - parameters are fixed depending on the data available. Can also list vector of parameter names to fix at their starting values (use param_adjust and val_adjust to set these adjustments)}

\item{C_opt}{default=0, NO catch data available. Copt=1 means the catch is in numbers, Copt2 means the catch is in weight.}

\item{Sel0}{input 1 means selectivity at age 0 fish is restrained to 1e-20; value of 0 means selectivity at age-0 fish can be larger following logistic curve}

\item{LFdist}{likelihood distribution for length composition data, default=0 for multinomial, alternate=1 for dirichlet-multinomial}
}
\value{
List, a tagged list of Data, Parameters, Random, Map
}
\description{
\code{format_input} Formats data, parameters, random effects, and mapped parameters for TMB input
}

