% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_input.R
\name{format_input}
\alias{format_input}
\title{TMB input formatting}
\usage{
format_input(input, data_avail, Fpen, SigRpen, SigRprior, est_sigma, f_startval,
  fix_param = FALSE, fix_param_t = FALSE, C_opt = 0, LFdist, S_l_input,
  theta_type, randomR)
}
\arguments{
\item{input}{tagged list of LIME inputs. Output from create_inputs.}

\item{data_avail}{types of data included, must at least include LCX where X is the number of years of length composition data. May also include "Catch" or "Index" separated by underscore. For example, "LC10", "Catch_LC1", "Index_Catch_LC20".}

\item{Fpen}{penalty on fishing mortality 0= off, 1=on}

\item{SigRpen}{penalty on sigmaR, 0=off, 1=on}

\item{SigRprior}{vector with prior info for sigmaR penalty, first term is the mean and second term is the standard deviation}

\item{est_sigma}{list of variance parameters to estimate, must match parameter names: log_sigma_R, log_sigma_C, log_sigma_I, log_CV_L, log_sigma_F}

\item{f_startval}{default=NULL and F starting values are at 0 for all years. Can also specify vector of F starting values for all years to be modeled (can start at truth for debugging).}

\item{fix_param}{default=FALSE - parameters are fixed depending on the data available. Can also list vector of parameter names to fix at their starting values (use param_adjust and val_adjust to set these adjustments)}

\item{fix_param_t}{default=FALSE - fix certain parameters in time series (e.g. fishing mortality, recruitment deviations) list with first item the name of the parameter and second item the numbers in the time series to be fixed.}

\item{C_opt}{default=0, NO catch data available. Copt=1 means the catch is in numbers, Copt2 means the catch is in weight.}

\item{LFdist}{likelihood distribution for length composition data, default=0 for multinomial, alternate=1 for dirichlet-multinomial}

\item{S_l_input}{input a vector specifying selectivity-at-length, or set less than 0 to use 1-parameter logistic function for selectivity}

\item{theta_type}{if 0, estimate annual theta; if 1, estimate single theta for all years of length comp}

\item{randomR}{default = TRUE, estimate recruitment as a random effect; if FALSE, turn off random effect on recruitment (do not derive deviations)}
}
\value{
List, a tagged list of Data, Parameters, Random, Map
}
\description{
\code{format_input} Formats data, parameters, random effects, and mapped parameters for TMB input
}
