% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIME_functions.R
\name{formatDataList}
\alias{formatDataList}
\title{Formatting real data for specific species for TMB model}
\usage{
formatDataList(species, data_dir)
}
\arguments{
\item{species}{species code, "SIGSUT" or "CRSNAP"}

\item{data_dir}{Directory to look for data files}
}
\value{
List, a tagged list of data to input into TMB model
}
\description{
\code{formatDataList} Reads data files from directory for specific species and gets the data into a form to input to the TMB model
}

