% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data for simulation testing}
\usage{
generate_data(modpath, data_avail, itervec, Fdynamics, Rdynamics, lh,
  pool = TRUE, Nyears, comp_sample, rewrite = TRUE, mismatch = FALSE,
  init_depl, derive_quants = FALSE)
}
\arguments{
\item{modpath}{directory to save generated data}

\item{data_avail}{types of data included, must at least include LCX where X is the number of years of length composition data. May also include Catch or Index separated by underscore. For example, LC10, Catch_LC1, Index_Catch_LC20.}

\item{itervec}{number of iterations of data to generate}

\item{Fdynamics}{Specify name of pattern of fishing mortality dynamics, Constant, Endogenous, Ramp, Increasing, or None}

\item{Rdynamics}{Specify name of pattern of recruitment dynamics, Constant, Pulsed, Pulsed_up, or BH}

\item{lh}{list of life history information to feed to population simulation function, output from create_lh_list}

\item{pool}{if nseasons (in life history list) is greater than one, pool the generated data into annual time steps, or leave at the season-level? default=TRUE, FALSE will generate shorter time step life history info, mean length}

\item{Nyears}{number of years to simulate}

\item{comp_sample}{vector with sample sizes of length composition data each year}

\item{rewrite}{TRUE will re-run OM and observation model. FALSE will skip if it's already written in directory.}

\item{mismatch}{default=FALSE, if TRUE, catch and index overlap with length comp only 1 year}

\item{init_depl}{default=0.4, can specify a different value or 2 values that indicate range from which to choose them}

\item{derive_quants}{default=FALSE (takes longer to run), can set to TRUE to output additional derived quantities.}
}
\value{
print how many iterations were written into the model directory
}
\description{
\code{generate_data} Generates data from the operating model for use in simulation testing
}
