\name{LIME-package}
\alias{LIME-package}
\alias{LIME}
\docType{package}
\title{
\packageTitle{LIME}
}
\description{
\packageDescription{LIME}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LIME}
\packageIndices{LIME}
Accounts for variation in fishing mortality and recruitment using a mixed effects approach; requires at least one year of length composition data and assumptions about biological parameters, but flexible to include more years of length composition data and/or catch and an abundance index time series.
}
\author{
\packageAuthor{LIME}

Maintainer: \packageMaintainer{LIME}
}
\keyword{ package }

