\name{LIME-package}
\alias{LIME-package}
\alias{LIME}
\docType{package}
\title{
\packageTitle{LIME}
}
\description{
\packageDescription{LIME}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LIME}
\packageIndices{LIME}
}
\author{
\packageAuthor{LIME}

Maintainer: \packageMaintainer{LIME}
}
\references{ Manuscript for LIME method in prep, email Merrill Rudd (merrillrudd@gmail.com) for details. }
\keyword{ package }
