% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_LCfits.R
\name{plot_LCfits}
\alias{plot_LCfits}
\title{plot length composition data and model fits}
\usage{
plot_LCfits(Inputs, Inputs2 = NULL, Inputs3 = NULL, Inputs4 = NULL,
  Report = NULL, Report2 = NULL, LBSPR = NULL, true_lc_years = NULL,
  ylim = NULL, ML50 = NULL, SL50 = NULL, dim = NULL, n = FALSE)
}
\arguments{
\item{Inputs}{LIME input file, must have "LF"=length frequency and "T_yrs"=total vector of years}

\item{Inputs2}{optional LIME input file for length data comparison; default NULL}

\item{Inputs3}{optional LIME input file for length data comparison; default NULL}

\item{Inputs4}{optional LIME input file for length data comparison; default NULL}

\item{Report}{LIME report file; default NULL if only plotting length data}

\item{Report2}{optional LIME report file for model fit comparison; default NULL}

\item{LBSPR}{LBSPR results - must have pLF = probability of being harvested in a length bin; default NULL}

\item{true_lc_years}{true years of length comp data for plotting; default NULL}

\item{ylim}{ylim for plot; default NULL}

\item{ML50}{length at 50% maturity for comparison; default NULL}

\item{SL50}{length at 50% selectivity for comparison; default NULL}

\item{dim}{dimensions of plot; default NULL if not specified, will approximate best based on number of years to plot}

\item{n}{if TRUE, will display sample size of length comp data; default FALSE}
}
\value{
figure with length composition data and model fits if Report or LBSPR are specified
}
\description{
\code{plot_LCfits} plot length composition data with option for model fits from LIME and LB-SPR
}
