% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_LCfits.R
\name{plot_LCfits}
\alias{plot_LCfits}
\title{plot length composition data and model fits}
\usage{
plot_LCfits(LF_df = NULL, binwidth = 1, Inputs = NULL,
  Report = NULL, LBSPR = NULL, plot_fit = TRUE, n = FALSE,
  time_labels = NULL, plot_type = "density")
}
\arguments{
\item{LF_df}{list of length frequency matrices}

\item{binwidth}{bin width for plotting data with LF_df}

\item{Inputs}{LIME input file; default NULL if only plotting length data}

\item{Report}{LIME report file; default NULL if only plotting length data}

\item{LBSPR}{LBSPR results - must have pLF = probability of being harvested in a length bin; default NULL}

\item{plot_fit}{default = TRUE if Report file or LBSPR file is included, plot the model fits, otherwise plot data only}

\item{n}{if TRUE, will display sample size of length comp data; default FALSE}

\item{time_labels}{default=NULL, otherwise a vector of time series names}

\item{plot_type}{default = "density", alternate = "counts"}
}
\value{
figure with length composition data and model fits if Report or LBSPR are specified
}
\description{
\code{plot_LCfits} plot length composition data with option for model fits from LIME and LB-SPR
}
\author{
M.B. Rudd
}
