% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_LCfits.R
\name{plot_LCfits}
\alias{plot_LCfits}
\title{plot length composition data and model fits}
\usage{
plot_LCfits(LFlist, Inputs = NULL, Report = NULL, LBSPR = NULL,
  ylim = NULL, ML50 = NULL, SL50 = NULL, dim = NULL, n = FALSE)
}
\arguments{
\item{LFlist}{list of length frequency matrices}

\item{Inputs}{LIME input file; default NULL if only plotting length data}

\item{Report}{LIME report file; default NULL if only plotting length data}

\item{LBSPR}{LBSPR results - must have pLF = probability of being harvested in a length bin; default NULL}

\item{ylim}{ylim for plot; default NULL}

\item{ML50}{length at 50 percent maturity for comparison; default NULL}

\item{SL50}{length at 50 percent selectivity for comparison; default NULL}

\item{dim}{dimensions of plot; default NULL if not specified, will approximate best based on number of years to plot}

\item{n}{if TRUE, will display sample size of length comp data; default FALSE}
}
\value{
figure with length composition data and model fits if Report or LBSPR are specified
}
\description{
\code{plot_LCfits} plot length composition data with option for model fits from LIME and LB-SPR
}
\author{
M.B. Rudd
}

