% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_output.R
\name{plot_output}
\alias{plot_output}
\title{plot LIME output}
\usage{
plot_output(
  Inputs = NULL,
  Report = NULL,
  Sdreport = NULL,
  LBSPR = NULL,
  lh,
  true_years = NULL,
  True = NULL,
  plot = c("Fish", "Rec", "SPR", "ML", "SB", "Selex"),
  set_ylim = list(SPR = c(0, 1)),
  legend = FALSE
)
}
\arguments{
\item{Inputs}{LIME input file}

\item{Report}{LIME report file}

\item{Sdreport}{LIME standard error file}

\item{LBSPR}{LBSPR results - either straight from LBSPR output or LIME-edited}

\item{lh}{life history list}

\item{true_years}{vector of true years (in case all_years and lc_years are 1:20 instead of 1998:2017)}

\item{True}{default=NULL, possible to specify true list from generated data if simulation}

\item{plot}{options for plotting include "Fish"=fishing mortality, "Rec"=recruitment (LIME only), "SPR"=spawning potential ratio, "ML"=mean length (including observed values; LIME only), "SB"=spawning biomass (LIME only), "Selex"=selectivity-at-length}

\item{set_ylim}{list to set ylim for each of the 'plot' arguments}
}
\value{
figure with length composition data and model fits if Report or LBSPR are specified
}
\description{
\code{plot_output} plot output from LIME or LB-SPR
}
\author{
M.B. Rudd
}
