% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.letter.R
\name{print.letter}
\alias{print.letter}
\title{Print letter on a figure}
\usage{
\method{print}{letter}(label = "(a)", xy = c(0.1, 0.925), ...)
}
\arguments{
\item{label}{label to display on figure panel}

\item{xy}{x and y in terms of percentages of total axes, for plotting text}

\item{...}{other arguments to 'text'}
}
\value{
text on a figure panel
}
\description{
\code{print.letter} Print a letter to label a panel on a figure
}
\author{
T.A. Branch
}

