% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_LIME.R
\name{run_LIME}
\alias{run_LIME}
\title{run LIME model - simulation mode}
\usage{
run_LIME(modpath, write = TRUE, lh, input_data, est_sigma, data_avail,
  itervec = NULL, REML = FALSE, rewrite = TRUE, simulation = TRUE,
  param_adjust = FALSE, val_adjust = FALSE, f_true = FALSE,
  fix_param = FALSE, C_opt = 0, F_up = 10, LFdist = 0,
  derive_quants = FALSE)
}
\arguments{
\item{modpath}{model directory}

\item{write}{default=TRUE, write to modpath; if FALSE, return output as list}

\item{lh}{list of life history information, from create_lh_list}

\item{input_data}{tagged list of data inputs. Required: years = vector of years (true years or indices); LF = matrix of length frequency (years along rows and length bins along columns), obs_per_year = vector of sample size per year. Optional: I_t = vector of abundance index, named with years; C_t = vector of catch, named with years.}

\item{est_sigma}{list of variance parameters to estimate, must match parameter names: log_sigma_R, log_sigma_C, log_sigma_I, log_CV_L, log_sigma_F}

\item{data_avail}{types of data included, must at least include LCX where X is the number of years of length composition data. May also include Catch or Index separated by underscore. For example, LC10, Catch_LC1, Index_Catch_LC20.}

\item{itervec}{number of datasets to generate in a simulation study. default=NULL for real stock assessment application.}

\item{REML}{default off (FALSE)}

\item{rewrite}{default=TRUE; if results already exist in the directory, should we rewrite them? TRUE or FALSE}

\item{simulation}{is this a simulation? default TRUE, FALSE means you are using real data (can set itervec=NULL)}

\item{param_adjust}{character or vector of parameter names to change input values}

\item{val_adjust}{number or vector of numbers for corresponding parameter value changes}

\item{f_true}{default=FALSE will make starting logF values =0; change to true and will use true values from simulation}

\item{fix_param}{default=FALSE - parameters are fixed depending on the data available. Can also list vector of parameter names to fix at their starting values (use param_adjust and val_adjust to set these adjustments)}

\item{C_opt}{default=0, if no catch data is available, set to 0. If catch is in numbers, set to 1. if catch is in biomass, set to 2.}

\item{F_up}{upper bound of fishing mortality estimate; default=5}

\item{LFdist}{likelihood distribution for length composition data, default=0 for multinomial, alternate=1 for dirichlet-multinomial}

\item{derive_quants}{default=FALSE (takes longer to run), can set to TRUE to output additional derived quantities.}
}
\value{
prints how many iterations were run in model directory
}
\description{
\code{run_LIME} run length-based integrated mixed-effects model with generated data
}
