% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runstack.R
\name{runstack}
\alias{runstack}
\title{Predictive stacking}
\usage{
runstack(savedir, lh, nodes, param, mean, cov, modname, data_avail = "LC",
  max_gradient = 0.001, C_type = 0, LFdist = 1, rewrite = TRUE,
  input_data = NULL, simulation = FALSE, iter = NULL, seed = NULL,
  Fscenario = NULL, model = "LIME")
}
\arguments{
\item{savedir}{directory to save results}

\item{lh}{life history list, with elements contained as output from LIME::create_lh_list}

\item{nodes}{matrix of nodes where each column is a different parameter and each row is a value from a distribution}

\item{param}{parameters (column names for nodes)}

\item{mean}{means of each parameter value,}

\item{cov}{covariance matrix across parameters}

\item{modname}{model name to save in directory}

\item{data_avail}{data available to model, default = "LC", can adjust to "Catch_LC", "Index_LC", or "Index_Catch_LC"}

\item{max_gradient}{maximum final gradient criterion, default = 0.001}

\item{C_type}{default = 0 (no catch data), 1=catch in numbers, 2= catch in biomass}

\item{LFdist}{default = 1 dirichlet-multinomial, 0=multinomial}

\item{rewrite}{rewrite results?}

\item{input_data}{for LIME use with real data (not simulated)}

\item{simulation}{is this a simulation? Default = FALSE}

\item{iter}{iteration of generated data}

\item{seed}{set seed}

\item{Fscenario}{fishing mortality scenario to generate data}

\item{model}{default="LIME", alternate = "LBSPR"}
}
\value{
prints how many iterations were run in model directory
}
\description{
\code{runstack} run predicted stacking for life history parameters using LIME
}
\author{
M.B. Rudd
}

