% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lime.R
\name{select_tree}
\alias{select_tree}
\title{Tree model for feature selection}
\usage{
select_tree(x, y, weights, n_features)
}
\arguments{
\item{x}{the data as a sparse matrix}

\item{y}{the labels}
}
\description{
Tree model for feature selection
}
