% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lime.R
\name{select_tree}
\alias{select_tree}
\title{Tree model for feature selection
Based on the latest XGBoost version.
May require the Drat package because of a bug in old version of xgb.model.dt.tree}
\usage{
select_tree(x, y, weights, n_features)
}
\arguments{
\item{x}{the data as a sparse matrix}

\item{y}{the labels}

\item{weights}{distance of the sample with the original datum}

\item{n_features}{number of features to take}
}
\description{
Tree model for feature selection
Based on the latest XGBoost version.
May require the Drat package because of a bug in old version of xgb.model.dt.tree
}
