% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_pop}
\alias{sim_pop}
\title{Operating model}
\usage{
sim_pop(lh, Nyears, pool, Fdynamics, Rdynamics, Nyears_comp, comp_sample,
  init_depl = 0.99, seed, sample_type = "catch", mgt_type = "F",
  fleet_proportions, nburn = 20)
}
\arguments{
\item{lh}{list of life history information, from create_lh_list}

\item{Nyears}{number of years to simulate}

\item{pool}{if nseasons (in life history list) is greater than one, pool the generated data into annual time steps, or leave at the season-level? FALSE will generate shorter time step life history info, mean length}

\item{Fdynamics}{Specify name of pattern of fishing mortality dynamics, Constant, Endogenous, Oneway, or None. Input number to project forward using a specific F.}

\item{Rdynamics}{Specify name of pattern of recruitment dynamics, Constant, Pulsed, Pulsed_up, or BH}

\item{Nyears_comp}{number of years of length composition data}

\item{comp_sample}{sample size of length composition data annually}

\item{init_depl}{initial depletion on which to calculate F1; default = 0.99}

\item{seed}{set seed for generating stochastic time series}

\item{sample_type}{a character vector specifying if the length comps are sampled from the 'catch' (default) or from the population}

\item{mgt_type}{removals based on F (default) or catch}

\item{fleet_proportions}{vector specifying the relative size of each fleet in terms of fishing pressure. must have length = nfleets and sum to 1.}

\item{nburn}{number of years of simulation burn in, default=20}
}
\value{
named list of attributes of true population/data
}
\description{
\code{sim_pop} Age-converted-to-length-based operating model specifying true population dynamics
}
\author{
M.B. Rudd
}

