% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_pop}
\alias{sim_pop}
\title{Operating model}
\usage{
sim_pop(lh, Nyears, Fdynamics, Rdynamics, Nyears_comp, comp_sample, init_depl,
  nburn, seed, modname)
}
\arguments{
\item{lh}{list of life history information, from create_lh_list}

\item{Nyears}{number of years to simulate}

\item{Fdynamics}{Specify name of pattern of fishing mortality dynamics, Constant, Endogenous, Ramp, Increasing, or None}

\item{Rdynamics}{Specify name of pattern of recruitment dynamics, Constant, Pulsed, Pulsed_up, or BH}

\item{Nyears_comp}{number of years of length composition data}

\item{comp_sample}{vector with sample sizes of length composition data annually}

\item{init_depl}{initial depletion; if FALSE, will use F1 from lh list}

\item{nburn}{number of years of burn-in for operating model}

\item{seed}{set seed for generating stochastic time series}

\item{modname}{save model name for true dynamics in named list output}
}
\value{
named list of attributes of true population/data
}
\description{
\code{sim_pop} Age-converted-to-length-based operating model specifying true population dynamics
}

