% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapper_example_data.R
\docType{data}
\name{snapper}
\alias{snapper}
\title{Example data from Costa Rican spotted rose snapper.}
\format{An object of class \code{list} of length 36.}
\source{
{Conservation International}
}
\usage{
snapper
}
\description{
A dataset containing the length composition, biological information, and abundance index for Costa Rican spotted rose snapper.
}
\keyword{datasets}

